Re-factored ResetPasswordConfirm using the open-closed principle
The .post method in ResetPasswordConfirm was very bulky and much of the functionality inside the post method made more sense as a collection of smaller methods with a more precise task, because they could be re-defined for another developer's use-case (this is my case). Functionality is left unchanged but the steps can now be adjusted if a developer subclasses this view. The post method is also now much more readable.

- The token is only used to fetch the user from the DB, so we have the "get_user_from_token" method. 
- The purpose of the try-except block for the password validation is now clearer; isolated as a method, the method validates the password and converts the Django validation error into a DRF one
- We used a contextmanager to trigger signals instead of just writing the signals. This gives a child class more control over what to do before and after the password change, improves readability, and puts all signal triggering in one place.
- There's an actual "reset_password" method now.